Implement hardware functionality
Implemented use of a button (physical button attached to case for raspberry pi) that allows the user to start recording and stop recording their audio. Also implemented a red light (physical red light attached to case for raspberry pi) that is solid when the py is ready for button input, slowly blinking when the py is loading, and a quick 3-time blink when the py has started recording or when the py has sent an email.

Both implementations used the gpiozero library.